fix: 영화 목록 페이지 첫 로드 시 api 2번 호출되는 버그 해결
